feat: User configurable serial ports (#55)
This change adds automatic serial port monitoring using IOKit and
introduces a number of hacks to the custom branch of plptools to allow
ncpd to be stopped and started again within the same process. These
hacks are currently required due to the use of global state in ncpd—it
really expects to be run as a single standalone process and respawned
whenever any configuration changes. Ideally, state would be better
contained with support for something like SIGHUP to reload configuration
and restart.