feat: URL handler to allow one-click copy-to-device / install
This change adds a handler for `x-reconnect://install/` URLs. They accept a `path` parameter of a file URL (e.g., `file:///Users/jbmorley/installer.sis`) and will copy the specified file to the Psion. Longer-term this should kick-off a managed install in Reconnect. Nearer-term, it could perhaps launch the Installer on the Psion on transfer completion.