refactor: Create ReconnectCore package for sharing code between multiple targets
This is preparatory work for separating Reconnect out into two separate apps: a transient GUI app, and a resident Menu Bar app. The architecture of plptools makes this kind of separation fairly easy, but we still need to share some code between the apps in the form of this new core package.