refactor: Create ReconnectCore package for sharing code between multiple targets (#132)
This is preparatory work for separating Reconnect out into two separate
apps: a transient GUI app, and a resident Menu Bar app. The architecture
of plptools makes this kind of separation fairly easy, but we still need
to share some code between the apps in the form of this new core
package.

The user-facing goal of this separation is to address the issue of
multitasking focus in macOS when dynamically hiding and showing the Dock
icon.