docs: Add documentation section to the website and links to required hardware
This is a pipe cleaner change to add the a tiny bit of documentation to the website to prove the architecture and layout for adding future documentation. It introduces a custom Jekyll plugin that generates a file tree for the documentation section; ideally I would use a third-party plugin for this but I’ve been unable to find one.