fix: 📁 Support installing files to the root of drives (#223)
Psions treat existance checks for drive roots differently to regular
files and directoires meaning we were incorrectly reporting `C:\` as not
found, resulting in an attempt to create the path, and subsequent
failure as it already existed 🤦🏻‍♂️.

This change catches up to an updated version of plptools which includes
a new exists check. We adopt this to avoid having to get extended
attributes as a proxy for the exists check, and also explicitly query
drive info for drive-root paths to check they exist.