feat: 📷 Take screenshots of EPOC32 devices (#244)
This change introduces Psion-side tools (included as a pre-built
`screenshot.exe` binary and installer right now), a menu option to
install them, and a workflow for calling `screenshot.exe` remotely. It
also refactors the mechanism for converting files to make it possible to
ensure MBM files are always converted when transferring screenshots,
irrespective of the user's global preference.