fix: ⚙️ Improve connection reliability (#271)
This change moves serial device connection termination into a new
LaunchAgent (reconnectd) which matches the process lifecycle of the menu
bar app. With the introduction of reconnectd, this change allows easy
sharing of connection state with both the menu bar and browser apps, and
updates the app to respond accordingly and surface relevant state to the
user.

Includes a drive-by fix to add configuration options for baud rate, with
ports configured to automatic baud rate selection by default. This
effectively adds support for Series 3a and 3 devices.